Generalize the use of size_t in TimeIntegration

+++ library/LibUtilities/TimeIntegration/CNABTimeIntegrationScheme.h
@@ -54,7 +54,7 @@ namespace LibUtilities
 class CNABTimeIntegrationScheme : public TimeIntegrationSchemeGLM
 {
 public:
-    CNABTimeIntegrationScheme(std::string variant, unsigned int order,
+    CNABTimeIntegrationScheme(std::string variant, size_t order,
                               std::vector<NekDouble> freeParams)
         : TimeIntegrationSchemeGLM("", 2, freeParams)
     {
@@ -78,8 +78,7 @@ public:
     }
 
     static TimeIntegrationSchemeSharedPtr create(
-        std::string variant, unsigned int order,
-        std::vector<NekDouble> freeParams)
+        std::string variant, size_t order, std::vector<NekDouble> freeParams)
     {
         boost::ignore_unused(variant);
         boost::ignore_unused(order);
@@ -141,7 +140,7 @@ public:
         phase->m_numMultiStepValues         = 1;
         phase->m_numMultiStepImplicitDerivs = 1;
         phase->m_numMultiStepDerivs         = 2;
-        phase->m_timeLevelOffset = Array<OneD, unsigned int>(phase->m_numsteps);
+        phase->m_timeLevelOffset    = Array<OneD, size_t>(phase->m_numsteps);
         phase->m_timeLevelOffset[0] = 0;
         phase->m_timeLevelOffset[1] = 0;
         phase->m_timeLevelOffset[2] = 0;
